.px_execute.php と px-files が同階層にない場合に Notice レベルのエラーが発生する不具合を修正。
他、テスト項目追加、ドキュメント追記など。
