+++ phpunit.xml
@@ -1,7 +1,7 @@
 <phpunit>
 	<testsuites>
 		<testsuite>
-			<file>./px-files/tests/mainTest.php</file>
+			<file>./tests/mainTest.php</file>
 		</testsuite>
 	</testsuites>
 </phpunit>
