+++ tests/testdata/standard/.px_execute.php
@@ -1,0 +1,4 @@
+<?php
+chdir(__DIR__);
+@require_once( '../../../vendor/autoload.php' );
+new picklesFramework2\pickles('./px-files/');
