+++ tests/testdata/standard/px-files/config.php
@@ -5,7 +5,7 @@ return call_user_func( function(){
 	$conf = new stdClass;
 
 	// project
-	$conf->name = 'Px2-MultiTheme'; // サイト名
+	$conf->name = 'Px2-MultiTheme - test - standard'; // サイト名
 	$conf->domain = null; // ドメイン
 	$conf->path_controot = '/'; // コンテンツルートディレクトリ
 
@@ -113,9 +113,9 @@ return call_user_func( function(){
 			'cookie_theme_switch'=>'THEME',
 			'path_theme_collection'=>'./px-files/themes/',
 			'attr_bowl_name_by'=>'data-contents-area',
-			'default_theme_id'=>'pickles2',
+			'default_theme_id'=>'standard',
 			'options'=>array(
-				'pickles2'=>array(
+				'standard2'=>array(
 					'sample_param'=>'hoge'
 				)
 			)
