+++ tests/testdata/standard/px-files/themes/standard/default.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset="utf-8">
+        <title>standard - default.html</title>
+    </head>
+    <body>
+        <p>default.html of theme &quot;standard&quot;</p>
+<p><?php print ( $theme->get_option('sample_param')==null ? 'success' : 'FAILED' ); ?></p>
+
+    </body>
+</html>
