+++ tests/testdata/standard/px-files/themes/standard2/default.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset="utf-8">
+        <title>standard2 - default.html</title>
+    </head>
+    <body>
+        <p>default.html of theme &quot;standard2&quot;</p>
+<p><?php print ( $theme->get_option('sample_param')=='hoge' ? 'success' : 'FAILED' ); ?></p>
+
+    </body>
+</html>
